Move skinLimb_parser.py
